Slot update and clear should now work for multiple rows
Row name will be copied if destination row is empty, or row
to be completely replaced.

For HD44780 display, encoders 3 + 4 will have no effect,
the apply option will be visible in slot menu to apply changes
to multiple slots on a single row

+++ avr/cores/megacommand/MCL/MCLClipBoard.h
@@ -20,8 +20,8 @@ public:
   bool open();
   bool close();
 
-  bool copy(int col, int row, int w, int h);
-  bool paste(int col, int row);
+  bool copy(uint16_t col, uint16_t row, uint16_t w, uint16_t h);
+  bool paste(uint16_t col, uint16_t row);
 
 };
